Reset form before opening modal
Previously the form would only reset after adding a new book to the
library and closing the modal, but that fails to account for when form
input data persists after a page reload or after cancelling the new
book entry.

Resetting the form right before opening the modal is a simpler and more
encapsulated approach.